Move ComputeIntronRate into pipeline_utils and setup v2
